+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 
 
+## [0.4.6](https://github.com/Blobfolio/dowser/releases/tag/v0.4.6) - 2022-04-16
+
+### Added
+
+* `Extension::codegen` (compile-time helper)
+* `Extension::slice_ext`
+
+
+
 ## [0.4.5](https://github.com/Blobfolio/dowser/releases/tag/v0.4.5) - 2022-03-29
 
 ### Changed